+++ force.h
@@ -5,14 +5,14 @@
 
 
 typedef struct {
-	float x;
-	float y;
-	float z;
+	double x;
+	double y;
+	double z;
 
 } force;
 
 force * force_create();
-void force_init(force * f, float x, float y, float z);
+void force_init(force * f, double x, double y, double z);
 void force_invert(force * f);
 void force_destroy(force * f);
